index: shift about page to index page
The most hits to my website are on the index page. Unfortunately,
the earlier index page didn't paint any information as to what to
expect from me or from the website and thus it had high drop offs.

On checking out other sites like https://www.joelonsoftware.com,
https://www.kalzumeus.com, I notice that the first screen gives
the user something to be hooked on to and would be of good help
if I change.

The only reason I had my earlier index page with just a photo and
a few social links is because that's how most static site generators
theme work and while I am not using a theme, I sort of
gravitated towards that.